Allowing shared trigger cooldowns
