+++ cooldowns/__init__.py
@@ -2,7 +2,8 @@ from collections import namedtuple
 
 from .buckets import CooldownBucket, SlashBucket
 from .protocols import CooldownBucketProtocol
-from .cooldown import Cooldown, cooldown, shared_cooldown, TriggerCooldown
+from .cooldown import Cooldown, cooldown, shared_cooldown
+from .trigger_cooldown import TriggerCooldown
 from .static_cooldown import StaticCooldown, static_cooldown
 from .cooldown_times_per import CooldownTimesPer
 from .static_times_per import StaticTimesPer
