+++ docs/modules/objects/trigger_cooldown.rst
@@ -1,0 +1,9 @@
+TriggerCooldown Reference
+==================
+
+.. currentmodule:: cooldowns
+
+.. autoclass:: TriggerCooldown
+    :members:
+    :undoc-members:
+    :special-members: __init__
